fix(#14): email-service를 user-service 내부에 통합
