--- email-service/src/main/java/com/pawbridge/emailservice/EmailServiceApplication.java
@@ -1,13 +1,0 @@
-package com.pawbridge.emailservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EmailServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EmailServiceApplication.class, args);
-	}
-
-}
