--- email-service/src/main/java/com/pawbridge/emailservice/dto/response/EmailVerifiedResponse.java
@@ -1,5 +1,0 @@
-package com.pawbridge.emailservice.dto.response;
-
-public record EmailVerifiedResponse(
-        boolean verified
-) {}
