--- email-service/src/main/java/com/pawbridge/emailservice/exception/ExpiredCodeException.java
@@ -1,10 +1,0 @@
-package com.pawbridge.emailservice.exception;
-
-import com.pawbridge.emailservice.exception.common.ApplicationException;
-import com.pawbridge.emailservice.exception.common.ErrorCode;
-
-public class ExpiredCodeException extends ApplicationException {
-    public ExpiredCodeException() {
-        super(ErrorCode.EXPIRED_CODE);
-    }
-}
