--- email-service/src/main/java/com/pawbridge/emailservice/exception/TooManyAttemptsException.java
@@ -1,10 +1,0 @@
-package com.pawbridge.emailservice.exception;
-
-import com.pawbridge.emailservice.exception.common.ApplicationException;
-import com.pawbridge.emailservice.exception.common.ErrorCode;
-
-public class TooManyAttemptsException extends ApplicationException {
-    public TooManyAttemptsException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
