--- email-service/src/main/java/com/pawbridge/emailservice/exception/common/GlobalExceptionRestAdvice.java
@@ -1,119 +1,0 @@
-package com.pawbridge.emailservice.exception.common;
-
-import com.pawbridge.emailservice.util.ResponseDTO;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.DataAccessException;
-import org.springframework.data.redis.RedisConnectionFailureException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.validation.BindException;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.HttpRequestMethodNotSupportedException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.MissingServletRequestParameterException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.NoHandlerFoundException;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionRestAdvice {
-
-    @ExceptionHandler
-    public ResponseEntity<ResponseDTO<Void>> applicationException(ApplicationException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity
-                .status(e.getErrorCode().getHttpStatus())
-                .body(ResponseDTO.error(e.getErrorCode()));
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ResponseDTO<Void>> bindException(BindException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.BAD_REQUEST,
-                        e.getBindingResult().getAllErrors().get(0).getDefaultMessage()));
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ResponseDTO<Void>> dbException(DataAccessException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.INTERNAL_SERVER_ERROR, "서버 에러!"));
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ResponseDTO<Void>> serverException(RuntimeException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.INTERNAL_SERVER_ERROR, "서버 에러!"));
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ResponseDTO<Void>> handleValidationExceptions(
-            MethodArgumentNotValidException e) {
-
-        BindingResult bindingResult = e.getBindingResult();
-
-        List<String> fieldErrors = bindingResult.getFieldErrors()
-                .stream()
-                .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                .collect(Collectors.toList());
-
-        log.error(e.getMessage(), e);
-        String errorMessage = String.join(", ", fieldErrors);
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.BAD_REQUEST, errorMessage));
-    }
-
-    @ExceptionHandler(NoHandlerFoundException.class)
-    public ResponseEntity<ResponseDTO<Void>> handleNoHandlerFoundException(NoHandlerFoundException e) {
-        log.error("No handler found for {}: {}", e.getHttpMethod(), e.getRequestURL(), e);
-        return ResponseEntity
-                .status(HttpStatus.NOT_FOUND)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.NOT_FOUND, "유효하지 않은 엔드포인트입니다."));
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-    public ResponseEntity<ResponseDTO<Void>> handleHttpMessageNotReadableException(HttpMessageNotReadableException e) {
-        log.error("Bad Request Body: {}", e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.BAD_REQUEST, "요청 본문 형식이 올바르지 않습니다."));
-    }
-
-    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
-    public ResponseEntity<ResponseDTO<Void>> handleHttpRequestMethodNotSupportedException(HttpRequestMethodNotSupportedException e) {
-        log.error("Method Not Allowed: {}", e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.METHOD_NOT_ALLOWED)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.METHOD_NOT_ALLOWED, "지원하지 않는 HTTP 메소드입니다."));
-    }
-
-    @ExceptionHandler(MissingServletRequestParameterException.class)
-    public ResponseEntity<ResponseDTO<Void>> handleMissingServletRequestParameterException(MissingServletRequestParameterException e) {
-        log.error("Missing Request Parameter: {}", e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.BAD_REQUEST,
-                        String.format("필수 요청 파라미터 '%s'(이)가 누락되었습니다.", e.getParameterName())));
-    }
-
-    @ExceptionHandler(RedisConnectionFailureException.class)
-    public ResponseEntity<ResponseDTO<Void>> handleRedisConnectionFailure(RedisConnectionFailureException e) {
-        log.error("Redis connection failed: {}", e.getMessage(), e);
-        return ResponseEntity
-                .status(HttpStatus.SERVICE_UNAVAILABLE)
-                .body(ResponseDTO.errorWithMessage(HttpStatus.SERVICE_UNAVAILABLE,
-                        "인증 서비스에 일시적인 문제가 발생했습니다. 잠시 후 다시 시도해주세요."));
-    }
-}
