+++ user-service/build.gradle
@@ -32,9 +32,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-mail'
 	implementation 'org.springframework.boot:spring-boot-starter-security'
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-
-	// Spring Cloud Feign Client
-	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:4.2.0'
+	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
 	implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
 	implementation 'me.paulschwarz:spring-dotenv:4.0.0'
