+++ user-service/src/main/java/com/pawbridge/userservice/UserServiceApplication.java
@@ -2,14 +2,11 @@ package com.pawbridge.userservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 // Spring Data JPA Auditing을 전역적으로 활성화
 @EnableJpaAuditing
-// Feign Client 활성화
-@EnableFeignClients
 public class UserServiceApplication {
 
 	public static void main(String[] args) {
