+++ user-service/src/main/java/com/pawbridge/userservice/email/dto/request/SendVerificationCodeRequest.java
@@ -1,0 +1,12 @@
+package com.pawbridge.userservice.email.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
+public record SendVerificationCodeRequest(
+        @NotBlank(message = "이메일은 필수입니다.")
+        @Email(message = "올바른 이메일 형식이 아닙니다.")
+        @Size(max = 100, message = "이메일은 100자를 초과할 수 없습니다.")
+        String email
+) {}
