+++ user-service/src/main/java/com/pawbridge/userservice/email/dto/response/EmailVerifiedResponse.java
@@ -1,0 +1,5 @@
+package com.pawbridge.userservice.email.dto.response;
+
+public record EmailVerifiedResponse(
+        boolean verified
+) {}
