+++ user-service/src/main/java/com/pawbridge/userservice/exception/EmailSendingException.java
@@ -1,0 +1,15 @@
+package com.pawbridge.userservice.exception;
+
+import com.pawbridge.userservice.exception.common.ApplicationException;
+import com.pawbridge.userservice.exception.common.ErrorCode;
+
+public class EmailSendingException extends ApplicationException {
+    public EmailSendingException() {
+        super(ErrorCode.EMAIL_SENDING_FAILED);
+    }
+
+    public EmailSendingException(String message, Throwable cause) {
+        super(ErrorCode.EMAIL_SENDING_FAILED, message);
+        initCause(cause);
+    }
+}
