+++ user-service/src/main/java/com/pawbridge/userservice/exception/InvalidCodeException.java
@@ -1,0 +1,10 @@
+package com.pawbridge.userservice.exception;
+
+import com.pawbridge.userservice.exception.common.ApplicationException;
+import com.pawbridge.userservice.exception.common.ErrorCode;
+
+public class InvalidCodeException extends ApplicationException {
+    public InvalidCodeException() {
+        super(ErrorCode.INVALID_CODE);
+    }
+}
