+++ user-service/src/main/java/com/pawbridge/userservice/exception/TooManyAttemptsException.java
@@ -1,0 +1,10 @@
+package com.pawbridge.userservice.exception;
+
+import com.pawbridge.userservice.exception.common.ApplicationException;
+import com.pawbridge.userservice.exception.common.ErrorCode;
+
+public class TooManyAttemptsException extends ApplicationException {
+    public TooManyAttemptsException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
