+++ user-service/src/main/resources/application.yml
@@ -21,6 +21,29 @@ spring:
         show_sql: true
     show-sql: true
 
+  # Redis 설정 (email-service 통합)
+  data:
+    redis:
+      host: ${REDIS_HOST:localhost}
+      port: 6379
+
+  # Mail 설정 (email-service 통합)
+  mail:
+    host: smtp.gmail.com
+    port: 587
+    username: ${GOOGLE_EMAIL}
+    password: ${GOOGLE_EMAIL_SECRET_KEY}
+    properties:
+      mail:
+        smtp:
+          auth: true
+          starttls:
+            enable: true
+            required: true
+          connectiontimeout: 5000
+          timeout: 5000
+          writetimeout: 5000
+
   security:
     oauth2:
       client:
@@ -38,7 +61,3 @@ jwt:
 
 oauth2:
   redirect-uri: ${OAUTH2_REDIRECT_URI}
-
-# Email Service 설정
-email-service:
-  url: ${EMAIL_SERVICE_URL:http://localhost:8088}
