+++ user-service/src/main/resources/templates/email_verification.html
