+++ user-service/src/main/resources/templates/password_reset.html
